refactor: cjsToEsm.cjs - rewrite it to its equivalent in ECMAScript notation. and rename it to esm.mjs
